# controller/openapi.yaml
...
openapi: 3.0.3
info:
  description: Hasura Storage is amazing
  version: 1.0.0
  title: Hasura Storage
  license:
    name: LICENSE-TBD
    url: https://LICENSE-TBD
servers:
  - url: http://{domain}:{port}/v1
    description: The development API server
    variables:
      domain:
        default: localhost
      port:
        default: '8000'
security:
  - Authorization: []
  - X-Hasura-Admin-Secret: []
components:
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key to authorize requests.
    X-Hasura-Admin-Secret:
      type: apiKey
      name: X-Hasura-Admin-Secret
      in: header
      description: Hasura admin secret
  schemas:
...
    FileMetadata:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        size:
          type: number
        bucketId:
          type: string
        eTag:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isUploaded:
          type: boolean
        mimeType:
          type: string
        uploadedByUserId:
          type: string
        metadata:
          type: object
          additionalProperties: true
        objectKey:
          type: string
        chunkSize:
          type: number
        chunkCount:
          type: number
        uploadId:
          type: string
...
    UpdateFileMetadata:
      type: object
      properties:
        name:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
paths:
...
  /files/{id}:
    head:
      summary: Retrieve information about a file
      description: Retrieve information about a file
      tags:
        - storage
      security:
        - Authorization: []
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: if-match
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Match
          in: header
          schema:
            type: string
        - name: if-none-match
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match
          in: header
          schema:
            type: string
        - name: if-modified-since
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Modified-Since
          in: header
          schema:
            type: string
        - name: if-unmodified-since
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Unmodified-Since
          in: header
          schema:
            type: string
        - name: q
          description: Quality of the image. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
        - name: h
          description: Resize image up to h maintaining aspect ratio. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
        - name: w
          description: Resize image up to w maintaining aspect ratio. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
        - name: b
          description: Blur the image according to this sigma value. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
      responses:
        '200':
          description: File information gathered successfully
          headers:
            Cache-Control:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
              schema:
                type: string
            Content-Length:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length
              schema:
                type: number
            Content-Type:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
              schema:
                type: string
            Etag:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Etag
              schema:
                type: string
            Last-Modified:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified
              schema:
                type: string
        '304':
          description: |
            File hasn't been modified based on:
            - file modification time is older than If-Modified-Since
            - etag matches one of If-None-Match
          headers:
            Cache-Control:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
              schema:
                type: string
            Content-Length:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length
              schema:
                type: number
            Content-Type:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
              schema:
                type: string
            Etag:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Etag
              schema:
                type: string
            Last-Modified:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified
              schema:
                type: string
        '412':
          description: |
            Some of the conditions specified in the headers failed to match. For instance:
              - etag doesn't match one of If-Match
              - if-unmodified-since is false
          headers:
            Cache-Control:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
              schema:
                type: string
            Content-Length:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length
              schema:
                type: number
            Content-Type:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
              schema:
                type: string
            Etag:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Etag
              schema:
                type: string
            Last-Modified:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified
              schema:
                type: string
        default:
          description: Some error occurred
          headers:
            X-Error:
              description: An error message
              schema:
                type: string
    get:
      summary: Retrieve contents of file
      description: Retrieve contents of file
      tags:
        - storage
      security:
        - Authorization: []
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: if-match
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Match
          in: header
          schema:
            type: string
        - name: if-none-match
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match
          in: header
          schema:
            type: string
        - name: if-modified-since
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Modified-Since
          in: header
          schema:
            type: string
        - name: if-unmodified-since
          description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Unmodified-Since
          in: header
          schema:
            type: string
        - name: q
          description: Quality of the image. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
        - name: h
          description: Resize image up to h maintaining aspect ratio. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
        - name: w
          description: Resize image up to w maintaining aspect ratio. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
        - name: b
          description: Blur the image according to this sigma value. Only applies to jpeg, webp and png files
          in: query
          schema:
            type: number
      responses:
        '200':
          description: File gathered successfully
          headers:
            Cache-Control:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
              schema:
                type: string
            Content-Length:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length
              schema:
                type: number
            Content-Type:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
              schema:
                type: string
            Etag:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Etag
              schema:
                type: string
            Last-Modified:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified
              schema:
                type: string
          content:
            application/octet-stream: {}
        '304':
          description: |
            File hasn't been modified based on:
            - file modification time is older than If-Modified-Since
          headers:
            Cache-Control:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
              schema:
                type: string
            Content-Length:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length
              schema:
                type: number
            Content-Type:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
              schema:
                type: string
            Etag:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Etag
              schema:
                type: string
            Last-Modified:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified
              schema:
                type: string
        '412':
          description: |
            Some of the conditions specified in the headers failed to match. For instance:
              - etag doesn't match one of If-Match
              - etag matches one of If-None-Match
              - if-unmodified-since is false
          headers:
            Cache-Control:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
              schema:
                type: string
            Content-Length:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length
              schema:
                type: number
            Content-Type:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
              schema:
                type: string
            Etag:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Etag
              schema:
                type: string
            Last-Modified:
              description: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified
              schema:
                type: string
        default:
          description: Some error occurred
          headers:
            X-Error:
              description: An error message
              schema:
                type: string
    put:
      summary: Replace an existing file with a new one
      description: |
        Replace the contents of an existing file. The workflow to replace a file is as follows:
          1. Set isUploaded flag to false
          2. Replace contents of the file in the storage backend
          3. Update file metadata (size, mime-type, isUploaded, etc)

        Each individual state is atomic but if a step fails, previous steps won't be undone
      tags:
        - storage
      security:
        - Authorization: []
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                metadata:
                  $ref: '#/components/schemas/UpdateFileMetadata'
                file:
                  description: New contents of the file to upload.
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '200':
          description: File was uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadata'
        default:
          description: En error occured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a file
      description: Delete a file
      tags:
        - storage
      security:
        - Authorization: []
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: File was deleted successfully
        default:
          description: En error occured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...